+++ FlOpEDT/TTapp/constraintHandler/specific_constraints/instructorConstraint.py
@@ -1,0 +1,14 @@
+from TTapp.constraintHandler.constraint import Constraint
+from TTapp.constraintHandler.constraint_type import ConstraintType
+
+
+class InstructorConstraint(Constraint):
+    def __init__(self, constraint_type, slot, course):
+        self.slot = slot
+        self.course = course
+        Constraint.__init__(self, constraint_type=constraint_type, courses=[course], slots=[slot])
+
+    def get_summary_format(self):
+        output = "\tLes cours suivants:\n%s\tdoivent avoir un professeur\n"
+        dimensions = ["courses"]
+        return output, dimensions