+++ CHANGELOG.md
@@ -84,3 +84,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [0.1] - 2018-11-06
 ### Added
 - Initial features
+
+### Added
+- TTApp/constraintHandler : contains all code + documentation allowing to write files in logs/ 
+that explain the infaisability by the solver of a set of constraints